[Meru AB] Bold version of assistant usage message
Fixes #3865
